% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bin_search}
\alias{bin_search}
\title{Find a value in a list using binary search
Wrapper for bin_search_rec}
\usage{
bin_search(l, val, start = NULL, end = NULL)
}
\arguments{
\item{l}{A list of values}

\item{val}{A value that is looked for}
}
\value{
Postition - 1 at which \code{val} must be in inserted to keep \code{l} sorted. \code{0} if \code{val} is
smaller than the first element of \code{l}, \code{length(l)} if it is greater than the last one
}
\description{
Find a value in a list using binary search
Wrapper for bin_search_rec
}
\examples{
bin_search(1:10, 2)
bin_search(1:10, 0)
bin_search(1:10, 2.5)
}
