% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.bin_search_rec}
\alias{.bin_search_rec}
\title{Binary search helper function
use bin_search instead}
\usage{
.bin_search_rec(l, val, start, end)
}
\description{
Binary search helper function
use bin_search instead
}
